Move officerCardData to separate officer-data.ts file
