+++ src/app/about/page.tsx
@@ -1,92 +1,8 @@
 import Image from "next/image";
 import styles from "./About.module.css";
 
-import OfficerCard, { OfficerCardProps } from "./officer-card";
-
-const officerCardData: OfficerCardProps[] = [
-  {
-    position: "President",
-    image: "public/officers/Yonas_Bahre.jpg",
-    name: "Yonas Bahre",
-    socialLinksData: [
-      {
-        name: "LinkedIn",
-        url: "https://www.linkedin.com/in/yonasbahre/",
-        icon: "/linkedin_logo.png",
-      },
-    ],
-  },
-  {
-    position: "Vice President",
-    image: "public/officers/Michael_Hayworth.jpg",
-    name: "Michael Hayworth",
-    socialLinksData: [
-      {
-        name: "LinkedIn",
-        url: "https://www.linkedin.com/in/michaeldhayworth/",
-        icon: "/linkedin_logo.png",
-      },
-    ],
-  },
-  {
-    position: "Treasurer",
-    image: "public/officers/Stephen_Coomes.jpg",
-    name: "Stephen Coomes",
-    socialLinksData: [
-      {
-        name: "LinkedIn",
-        url: "https://www.linkedin.com/in/stephen-coomes-8a4885221/",
-        icon: "/linkedin_logo.png",
-      },
-    ],
-  },
-  {
-    position: "Outreach Officer",
-    image: "public/officers/Angel_Lopez.png",
-    name: "Angel Lopez",
-    socialLinksData: [
-      {
-        name: "GitHub",
-        url: "https://github.com/angel1254mc",
-        icon: "/github_logo.svg",
-      },
-      {
-        name: "LinkedIn",
-        url: "https://www.linkedin.com/in/angel1254/",
-        icon: "/linkedin_logo.png",
-      },
-      {
-        name: "Website",
-        url: "https://www.angel1254.com/",
-        icon: "/web_logo.svg",
-      },
-    ],
-  },
-  {
-    position: "Program Officer",
-    image: "public/officers/Param_Gupta.jpg",
-    name: "Param Gupta",
-    socialLinksData: [
-      {
-        name: "LinkedIn",
-        url: "https://www.linkedin.com/in/paramg/",
-        icon: "/linkedin_logo.png",
-      },
-    ],
-  },
-  {
-    position: "Involvement Officer",
-    image: "public/officers/CJ_Weir.jpg",
-    name: "CJ Weir",
-    socialLinksData: [
-      {
-        name: "LinkedIn",
-        url: "https://www.linkedin.com/in/cj-weir/",
-        icon: "/linkedin_logo.png",
-      },
-    ],
-  },
-];
+import OfficerCard from "./officer-card";
+import officerCardData from "./officer-data";
 
 export default function About() {
   const officerCards = officerCardData.map((cardData) => (
